Remodularized mfe static methods
