+++ python/pymfe/mfe.py
@@ -1,7 +1,6 @@
 """Main module for extracting metafeatures from datasets."""
 import typing as t
 import collections
-import warnings
 
 import numpy as np
 
@@ -22,16 +21,15 @@ class MFE:
         splits (:obj:`Iterable`): Iterable/generator for K-Fold Cross
             Validation train and test indexes splits.
 
-        groups (:obj:`Sequence`): tuple object containing fitted me-
-            tafeature groups loaded in model at instantiation.
+        groups (:obj:`tuple` of :obj:`str`): tuple object containing fitted
+            metafeature groups loaded in model at instantiation.
 
-        features (:obj:`Sequence`): tuple object containing loaded
-            metafeature-extraction method names available for meta-
-            feature extraction, from selected metafeatures groups and
-            features listed at instantiation
+        features (:obj:`tuple` of :obj:`str`): contains loaded metafeature-
+            extraction method names available for metafeature extraction, from
+            selected metafeatures groups and features listed at instantiation
 
-        summary (:obj:`Sequence`): tuple object which contains summary
-            functions names for features summarization.
+        summary (:obj:`tuple` of :obj:`str`): tuple object which contains sum-
+            mary functions names for features summarization.
     """
     # pylint: disable=R0902
 
@@ -144,193 +142,6 @@ class MFE:
         self._custom_args_ft = None  # type: t.Optional[t.Dict[str, t.Any]]
         self._custom_args_sum = None  # type: t.Optional[t.Dict[str, t.Any]]
 
-    @staticmethod
-    def _summarize(features: t.Union[np.ndarray, t.Sequence],
-                   callable_sum: t.Callable,
-                   callable_args: t.Optional[t.Dict[str, t.Any]] = None,
-                   remove_nan: bool = True,
-                   ) -> t.Union[t.Sequence, _internal.TypeNumeric]:
-        """Returns feature summarized by `callable_sum`.
-
-        Args:
-            features (:obj:`Sequence` of numerics): Sequence containing values
-                to summarize.
-
-            callable_sum (:obj:`Callable`): Callable of the method which im-
-                plements the desired summary function.
-
-            callable_args (:obj:`Dict`, optional): arguments to the summary
-                function. The expected dictionary format is the following:
-                {`argument_name`: value}. In order to know the summary func-
-                tion arguments you need to check out the documentation of
-                the method which implements it.
-
-            remove_nan (:obj:`bool`, optional): check and remove all elements
-                in `features` which are not numeric. Note that :obj:`np.inf`
-                is still considered numeric (:obj:`float` type).
-
-        Returns:
-            float: value of summarized feature values if possible. May
-            return :obj:`np.nan` if summary function call invokes TypeError.
-
-        Raises:
-            AttributeError: if `callable_sum` is invalid.
-            TypeError: if `features`  is not a sequence.
-        """
-        processed_feat = np.array(features)
-
-        if remove_nan:
-            numeric_vals = list(map(_internal.isnumeric, features))
-            processed_feat = processed_feat[numeric_vals]
-            processed_feat = processed_feat.astype(np.float32)
-
-        if callable_args is None:
-            callable_args = {}
-
-        try:
-            metafeature = callable_sum(processed_feat, **callable_args)
-
-        except TypeError:
-            metafeature = np.nan
-
-        return metafeature
-
-    @staticmethod
-    def _get_feat_value(method_name: str,
-                        method_args: t.Dict[str, t.Any],
-                        method_callable: t.Callable,
-                        suppress_warnings: bool = False
-                        ) -> t.Union[_internal.TypeNumeric, np.ndarray]:
-        """Extract feat. from `method_callable` with `method_args` as args.
-
-        Args:
-            method_name (:obj:`str`): name of the feature-extraction method
-                to be invoked.
-
-            method_args (:obj:`Dic`): arguments of method to be invoked. The
-                expected format of the arguments is {`argument_name`: value}.
-                In order to know the method arguments available, you need to
-                check its documentation.
-
-            method_callable(:obj:`Callable`): callable of the feature-extra-
-                ction method.
-
-            suppress_warnings(:obj:`bool`): if True, all warnings invoked whi-
-                before invoking the method (or after) will be ignored. The me-
-                thod itself may still invoke warnings.
-
-        Returns:
-            numeric or array: return value of the feature-extraction method.
-
-        Raises:
-            AttributeError: if `method_callable` is not valid.
-        """
-
-        try:
-            features = method_callable(**method_args)
-
-        except TypeError as type_e:
-            if not suppress_warnings:
-                warnings.warn(
-                    "Error extracting {0}: \n{1}.\nWill set it "
-                    "as 'np.nan' for all summary functions.".format(
-                        method_name, repr(type_e)), RuntimeWarning)
-
-            features = np.nan
-
-        return features
-
-    @staticmethod
-    def _build_mtd_args(
-            method_name: str,
-            method_args: t.Iterable[str],
-            inner_custom_args: t.Optional[t.Dict[str, t.Any]] = None,
-            user_custom_args: t.Optional[t.Dict[str, t.Any]] = None,
-            suppress_warnings: bool = False) -> t.Dict[str, t.Any]:
-        """Build a `kwargs` (:obj:`Dict`) for a feature-extraction :obj:`Callable`.
-
-        Args:
-            method_name (:obj:`str`): name of the method.
-
-            method_args (:obj:`Iterable` of :obj:`str`): Iterable containing
-                the name of all arguments of the callable.
-
-            inner_custom_args (:obj:`Dict`, optional): custom arguments for
-                inner usage, for example, to pass ``X``, ``y`` or other user-
-                independent arguments necessary for the callable. The expected
-                format of this dict is {`argument_name`: value}.
-
-            user_custom_args (:obj:`Dict`, optional): assumes the same model
-                as the dict above, but this one is dedicated to keep user-dep-
-                endent arguments for method callable, for example, number of
-                bins of a histogram-like metafeature or degrees of freedom of
-                a standard deviation-related metafeature. The name of the ar-
-                guments must be verified in its correspondent method documen-
-                tation.
-
-            suppress_warnings(:obj:`bool`, optional): if True, will not show
-                any warnings about unknown callable parameters.
-
-        Returns:
-            dict: a ready-to-use `kwargs` for the correspondent callable. The
-                format is {`argument_name`: value}.
-        """
-
-        if user_custom_args is None:
-            user_custom_args = {}
-
-        if inner_custom_args is None:
-            inner_custom_args = {}
-
-        combined_args = {
-            **user_custom_args,
-            **inner_custom_args,
-        }
-
-        callable_args = {
-            custom_arg: combined_args[custom_arg]
-            for custom_arg in combined_args if custom_arg in method_args
-        }
-
-        if not suppress_warnings:
-            unknown_arg_set = (unknown_arg
-                               for unknown_arg in user_custom_args.keys()
-                               if unknown_arg not in method_args
-                               )  # type: t.Generator[str, None, None]
-
-            for unknown_arg in unknown_arg_set:
-                warnings.warn(
-                    'Unknown argument "{0}" for method "{1}".'.format(
-                        unknown_arg, method_name), UserWarning)
-
-        return callable_args
-
-    @staticmethod
-    def _check_summary_warnings(
-            value: t.Union[_internal.TypeNumeric, t.Sequence, np.ndarray],
-            name_feature: str,
-            name_summary: str) -> None:
-        """Check if there is :obj:`np.nan` within summarized values.
-
-        Args:
-            value (numeric or :obj:`Sequence`): summarized values.
-
-            name_feature (:obj:`str`): name of the feature-extraction
-                method used to generate the values which was summarized.
-
-            name_summary (:obj:`str`): name of the summary method
-                used to produce `value`.
-        """
-
-        if not isinstance(value, collections.Iterable):
-            value = [value]
-
-        if any(np.isnan(value)):
-            warnings.warn(
-                "Failed to summarize {0} with {1}. "
-                "(generated NaN).".format(name_feature,
-                                          name_summary), RuntimeWarning)
-
     def _call_summary_methods(
                 self,
                 feature_values: t.Sequence[_internal.TypeNumeric],
@@ -339,7 +150,7 @@ class MFE:
                 suppress_warnings: bool = False,
                 **kwargs
                 ) -> t.Tuple[t.List[str], t.List[t.Union[float, t.Sequence]]]:
-        r"""Invoke summary functions loaded in model on given feature values.
+        """Invoke summary functions loaded in model on given feature values.
 
         Args:
             feature_values (:obj:`Sequence` of numerics): sequence containing
@@ -375,21 +186,21 @@ class MFE:
 
         for sm_mtd_name, sm_mtd_callable, sm_mtd_args in self._metadata_mtd_sm:
 
-            sm_mtd_args_pack = MFE._build_mtd_args(
+            sm_mtd_args_pack = _internal.build_mtd_kwargs(
                 method_name=sm_mtd_name,
                 method_args=sm_mtd_args,
                 user_custom_args=kwargs.get(sm_mtd_name),
                 inner_custom_args=self._custom_args_sum,
                 suppress_warnings=suppress_warnings)
 
-            summarized_val = MFE._summarize(
+            summarized_val = _internal.summarize(
                 features=feature_values,
                 callable_sum=sm_mtd_callable,
                 callable_args=sm_mtd_args_pack,
                 remove_nan=remove_nan)
 
             if not suppress_warnings:
-                MFE._check_summary_warnings(
+                _internal.check_summary_warnings(
                     value=summarized_val,
                     name_feature=feature_name,
                     name_summary=sm_mtd_name)
@@ -503,7 +314,7 @@ class MFE:
                 ([`attr_ent.mean`, `attr_ent.sd`], [0.983459, 0.344361]) is
                 the return value for the feature `attr_end` summarized by
                 both `mean` and `sd` (standard deviation), giving the values
-                0.983469 and 0.344361, respectively.
+                `0.983469` and `0.344361`, respectively.
 
         Raises:
             TypeError: if calling `extract(...)` method before `fit(...)`
@@ -524,14 +335,14 @@ class MFE:
 
             ft_name_without_prefix = _internal.remove_mtd_prefix(ft_mtd_name)
 
-            ft_mtd_args_pack = MFE._build_mtd_args(
+            ft_mtd_args_pack = _internal.build_mtd_kwargs(
                 method_name=ft_name_without_prefix,
                 method_args=ft_mtd_args,
                 user_custom_args=kwargs.get(ft_name_without_prefix),
                 inner_custom_args=self._custom_args_ft,
                 suppress_warnings=suppress_warnings)
 
-            features = MFE._get_feat_value(
+            features = _internal.get_feat_value(
                 ft_mtd_name,
                 ft_mtd_args_pack,
                 ft_mtd_callable,
